@@ -179,20 +172,20 @@
     "redux-thunk": "^2.1.0",
     "request": "^2.79.0",
     "sass-lint": "^1.10.2",
-    "sass-loader": "^4.0.0",
+    "sass-loader": "^6.0.6",
     "saucelabs": "^1.4.0",
     "selenium-server": "^3.4.0",
     "semver": "^5.3.0",
     "sinon": "^3.2.1",
     "skin-deep": "^1.0.0",
-    "style-loader": "^0.13.1",
-    "svg-url-loader": "^1.1.0",
+    "style-loader": "^0.20.2",
+    "svg-url-loader": "^2.3.2",
     "tinyliquid": "^0.2.33",
-    "url-loader": "^0.5.7",
+    "url-loader": "^0.6.2",
     "uswds": "^1.4.2",
-    "webpack": "^2.4.1",
-    "webpack-manifest-plugin": "^1.1.0",
-    "webpack-md5-hash": "0.0.5",
+    "webpack": "^4.1.1",
+    "webpack-dev-server": "^3.1.1",
+    "webpack-manifest-plugin": "^2.0.0-rc.2",
     "whatwg-fetch": "^2.0.3",
     "winston": "^2.2.0"
   },
