@@ -23,20 +23,20 @@ function Docket({ ahead, total, form9Date, docketMonth, appealType, aod, front: 
   // This is the only part that's different between the two "normal" contents
   let yourPlaceText;
   if (frontOfDocket) {
-    yourPlaceText = <p>The Board is currently working on appeals from {docketMonthFormatted} or older. Your appeal is eligible to be assigned to a judge when it is ready for their review.</p>;
+    yourPlaceText = <p>The Board is currently reviewing appeals from {docketMonthFormatted} or older. Your appeal is eligible to be sent to a judge when it is ready for their review.</p>;
   } else {
-    yourPlaceText = <p>There are {total.toLocaleString()} appeals on the docket, not including Advanced on Docket and Court Remand appeals. Some of these appeals are not ready to be assigned to a judge. A judge will begin work on your appeal when it is among the oldest appeals that are ready for their review. The Board is currently working on appeals from {docketMonthFormatted} or older.</p>;
+    yourPlaceText = <p>There are {total.toLocaleString()} appeals on the docket, not including Advanced on Docket and Court Remand appeals. Some of these appeals are not ready to be sent to a judge. A judge will begin work on your appeal when it is among the oldest appeals that are ready for their review. The Board is currently reviewing appeals from {docketMonthFormatted} or older.</p>;
   }
 
   // Start with the basic content...
   let content = (
     <div>
-      <p>The Board of Veterans’ Appeals hears cases in the order they are received. When you completed a Form 9 in {form9DateFormatted}, your appeal was added to the Board’s docket, securing your spot in line.</p>
+      <p>The Board of Veterans’ Appeals reviews cases in the order they are received. When you completed a VA Form 9 in {form9DateFormatted}, your appeal was added to the Board’s docket, securing your spot in line.</p>
       {yourPlaceText}
       <DocketCard total={total} ahead={ahead}/>
-      <h2>Is there a way for my appeal to be decided more quickly?</h2>
-      <p>If you are suffering a serious illness or are in financial distress, or for other sufficient cause, you can apply to have your appeal <strong>Advanced on Docket</strong>. If you are older than 75, your appeal will have this status automatically. Advanced on Docket appeals are prioritized so that they are always at the front of the line.</p>
-      <p><a target="_blank" href="">Learn more about requesting Advanced on Docket status.</a></p>
+      <h2>Is there a way to prioritize my appeal?</h2>
+      <p>If you are suffering a serious illness or are in financial distress, or for other sufficient cause, you can apply to have your appeal <strong>Advanced on Docket</strong>. If you are older than 75, your appeal will receive this status automatically. Advanced on Docket appeals are prioritized so that they are always at the front of the line.</p>
+      <p><a target="_blank" href="/disability-benefits/claims-appeal/request-a-priority-review/">Learn more about requesting Advanced on Docket status.</a></p>
     </div>
   );
 
@@ -45,20 +45,20 @@ function Docket({ ahead, total, form9Date, docketMonth, appealType, aod, front: 
   if (aod) {
     content = (
       <div>
-        <p>Your appeal is Advanced on Docket. This could be because you or your representative requested this status, or because you are older than 75.</p>
-        <p>Advanced on Docket appeals are prioritized so that they are always at the front of the line. Your appeal will be assigned to a judge as soon as it is ready for their review.</p>
+        <p>Your appeal is Advanced on Docket. This could be because you are older than 75, because you are suffering a serious illness or are in financial distress, or for other sufficient cause.</p>
+        <p>Advanced on Docket appeals are prioritized so that they are always at the front of the line. Your appeal will be sent to a judge as soon as it is ready for their review.</p>
       </div>
     );
   }
 
   // This should override the aod section
   if (appealType === APPEAL_TYPES.postCavcRemand) {
-    content = <p>Your appeal was remanded by the Court of Appeals for Veterans’ Claims. Court Remand appeals are prioritized so that they are always at the front of the line. Your appeal will be assigned to a judge as soon as it is ready for their review.</p>;
+    content = <p>Your appeal was remanded by the Court of Appeals for Veterans’ Claims. Court Remand appeals are prioritized so that they are always at the front of the line. Your appeal will be sent to a judge as soon as it is ready for their review.</p>;
   }
 
   return (
     <div>
-      <h2>How long until a judge is ready to write your decision?</h2>
+      <h2>How long until a judge is ready for your appeal?</h2>
       {content}
     </div>
   );
