@@ -23,14 +23,22 @@ class Timeline extends React.Component {
     return `${first} - ${last}`;
   };
 
-  toggleExpanded = () => this.setState((prevState) => ({ expanded: !prevState.expanded }));
+  toggleExpanded = (e) => {
+    e.stopPropagation();
+    this.setState((prevState) => ({ expanded: !prevState.expanded }));
+  }
 
   render() {
     const { events, missingEvents } = this.props;
     let pastEventsList = [];
     if (events.length) {
       pastEventsList = events.map((event, index) => {
-        const { title, description, liClass } = getEventContent(event);
+        const content = getEventContent(event);
+        if (!content) {
+          return null;
+        }
+
+        const { title, description, liClass } = content;
         const date = formatDate(event.date);
         const hideSeparator = (index === events.length - 1);
         return (
@@ -42,7 +50,7 @@ class Timeline extends React.Component {
             liClass={liClass}
             hideSeparator={hideSeparator}/>
         );
-      });
+      }).filter(e => !!e); // Filter out the nulls
     }
 
     const downArrow = this.state.expanded ? <div className="down-arrow"/> : null;
