@@ -1,9 +1,12 @@
 import React from 'react';
+import classNames from 'classnames';
 
 export default function TitleField({ id, title }) {
   const isRoot = id === 'root__title';
 
-  return isRoot
-    ? <legend className="schemaform-block-title" id={id}>{title}</legend>
-    : <h5 className="schemaform-block-title" id={id}>{title}</h5>;
+  const classes = classNames('schemaform-block-title', {
+    'schemaform-block-subtitle': !isRoot
+  });
+
+  return <legend className={classes} id={id}>{title}</legend>;
 }
