@@ -2,6 +2,7 @@ import _ from '../../../common/utils/data-utils';
 
 import fullSchema526EZ from 'vets-json-schema/dist/21-526EZ-schema.json';
 
+import initialData from '../../../../../test/disability-benefits/526EZ/schema/initialData';
 
 import IntroductionPage from '../components/IntroductionPage';
 import ConfirmationPage from '../containers/ConfirmationPage';
@@ -12,9 +13,10 @@ import {
   evidenceTypesDescription,
   EvidenceTypeHelp,
   disabilityNameTitle,
+  vaMedicalRecordsIntro,
+  privateMedicalRecordsIntro,
   facilityDescription,
   treatmentView,
-  vaMedicalRecordsIntro
 } from '../helpers';
 
 const {
@@ -36,64 +38,6 @@ const treatmentsSchema = _.set('items.properties.treatment.properties',
     ],
     treatments.items.properties.treatment.properties
   ), treatments);
-
-const initialData = {
-  // For testing purposes only
-  disabilities: [
-    {
-      disability: { // Is this extra nesting necessary?
-        diagnosticText: 'PTSD',
-        decisionCode: 'Filler text', // Should this be a string?
-        // Is this supposed to be an array?
-        specialIssues: {
-          specialIssueCode: 'Filler text',
-          specialIssueName: 'Filler text'
-        },
-        ratedDisabilityId: '12345',
-        disabilityActionType: 'Filler text',
-        ratingDecisionId: '67890',
-        diagnosticCode: 'Filler text',
-        // Presumably, this should be an array...
-        secondaryDisabilities: [
-          {
-            diagnosticText: 'First secondary disability',
-            disabilityActionType: 'Filler text'
-          },
-          {
-            diagnosticText: 'Second secondary disability',
-            disabilityActionType: 'Filler text'
-          }
-        ]
-      }
-    },
-    {
-      disability: { // Is this extra nesting necessary?
-        diagnosticText: 'Second Disability',
-        decisionCode: 'Filler text', // Should this be a string?
-        // Is this supposed to be an array?
-        specialIssues: {
-          specialIssueCode: 'Filler text',
-          specialIssueName: 'Filler text'
-        },
-        ratedDisabilityId: '54321',
-        disabilityActionType: 'Filler text',
-        ratingDecisionId: '09876',
-        diagnosticCode: 'Filler text',
-        // Presumably, this should be an array...
-        secondaryDisabilities: [
-          {
-            diagnosticText: 'First secondary disability',
-            disabilityActionType: 'Filler text'
-          },
-          {
-            diagnosticText: 'Second secondary disability',
-            disabilityActionType: 'Filler text'
-          }
-        ]
-      }
-    }
-  ]
-};
 
 const formConfig = {
   urlPrefix: '/',
@@ -227,7 +171,7 @@ const formConfig = {
         },
         vaMedicalRecordsIntro: {
           title: '',
-          path: 'supporting-evidence/:index/va-medical-records',
+          path: 'supporting-evidence/:index/va-medical-records-intro',
           showPagePerItem: true,
           arrayPath: 'disabilities',
           depends: (formData, index) => _.get(`disabilities.${index}.view:vaMedicalRecords`, formData),
@@ -325,7 +269,33 @@ const formConfig = {
             }
           }
         },
-        // pageFive: {},
+        privateMedicalRecordsIntro: {
+          title: '',
+          path: 'supporting-evidence/:index/private-medical-records-intro',
+          showPagePerItem: true,
+          arrayPath: 'disabilities',
+          depends: (formData, index) => _.get(`disabilities.${index}.view:privateMedicalRecords`, formData),
+          uiSchema: {
+            disabilities: {
+              items: {
+                'ui:title': disabilityNameTitle,
+                'ui:description': privateMedicalRecordsIntro,
+              }
+            }
+          },
+          schema: {
+            type: 'object',
+            properties: {
+              disabilities: {
+                type: 'array',
+                items: {
+                  type: 'object',
+                  properties: {}
+                }
+              }
+            }
+          }
+        },
         // pageSix: {},
         // pageSeven: {},
         // pageEight: {},
