@@ -17,7 +17,7 @@ describe('<AppealsV2TabNav/>', () => {
     const wrapper = mount(<AppealsV2TabNav {...defaultProps}/>);
     const tabs = wrapper.find('li.appeals-tabs-item');
     const statusTab = wrapper.findWhere(c => c.type() === 'li' && c.text() === 'Status');
-    const detailTab = wrapper.findWhere(c => c.type() === 'li' && c.text() === 'Detail');
+    const detailTab = wrapper.findWhere(c => c.type() === 'li' && c.text() === 'Issues');
     expect(tabs.length).to.equal(2);
     expect(statusTab.length).to.equal(1);
     expect(detailTab.length).to.equal(1);
